Refactor e2e tests
This moves the common e2e code to a "helpers" module, while it renames
the "single" test case on to a more general e2e_test file.

This attempts to make the test code more readable, having the test code
be in one file, while the setup/helpers in another one.