Add a new CR ComplianceRemediation and a new controller
The ComplianceRemediation resource is used to contain MachineConfigs at
the moment, but in the future will contain also cluster-wide checks.

The ComplianceRemediation resources are process by a new controller
called complianceremediation_controller that looks for remediations
with apply=true only. Those are merged with remediations that are
already applied, transformed into a MachineConfig object which then
gets created in the cluster.

The MachineConfig object name is currently hard coded.