Enable using/pushing local image for e2e tests
This enables us to build and push the image to the openshift cluster
that we're working on before doing the end-to-end tests. With this, we
can easily test what we're working on instead of running the e2e tests
with the image that's in Quay.

This also introduces a way to not always trigger an image build when
running the e2e tests. This is done through the `E2E_SKIP_CONTAINER_PUSH`
flag, which will be taken into use if the value is something other than
`false`