Add a super simple ARF parser
OpenScap allows the administrator to select several different output
types. One of them is the ARF type:
    https://www.open-scap.org/features/scap-components/
ARF is a XML report that also contains metadata about the checks and
remediations. Therefore it is handy to use in order to see what checks
had failed and what remediations should be parsed.

This will probably be superseded in the future by a JSON parser, so
this ARF parser is not really much else than a hack.