Add interface for non-retriable errors
This creates an error wrapper called `NonRetriableCtrlError` which
represents an error that's fatal and should not be retried on the
controller.

This changes the logic in the controllers to use this specific
NonRetriableCtrlError as opposed to the wrapper that was implemented
originally.

It also adds the logic of retry/no-retry for errors in the
compliancesuite controller.