Use deep-compare when checking if all remediations for a scan are the same
We used to only check if the lenght of remediations for a scan (which
should map to a pool) is the same for all scan results. With this patch,
we also reflect and deep compare the remediations and their
machineconfig Specs and error out if one of the machines is out of step
with the rest.