compliancescan: Use a configMap as the script source for the openscap container, not the container's entrypoint
The openscap container uses a script baked into the container image that
represents the container's entry point. This patch instead uses a
configmap mounted into the container as an entrypoint and another
configmap with environment variables that fine-tune the entrypoint
script.

This has two advantages:
    1) All the logic is in one place, we no longer need to synchronize
    changes between the container image logic and the controller logic
    2) The openscap container image can be really dumb, just installing
    the scap scanner.

A follow-up patch will add another Dockerfile to the project so that we
can stop relying on the openscap container from a third party repository
and instead build another image from the operator repository.