Use urn:xccdf:fix:script:ignition remediations instead of pretend ignition in bash
There's been a bunch of commits recently to the ComplianceAsCode repo
that were either merged:
    https://github.com/ComplianceAsCode/content/commit/85587c4159fa316b404e74b0b1c77ac642a3d2b0
or are in progress:
    https://github.com/ComplianceAsCode/content/pull/5142

Those commits adds a new type of remediation that uses the type URN
"urn:xccdf:fix:script:ignition". This commits updates our ARF parser to
look for remediations of this type and points to a new content image for
e2e tests that require remediations.