tests: Also wait for the nodes to be have the desired rendered MC
We found out that the fact that a pool contains a rendered MC config and
reports as ready does not guarantee that a machine is up and running
with that pool. Let's also check that the pools the machine had booted
up with by checking the annotations of the node objects themselves.