e2e: Fix predicates for unapplying MCs by split unApplyRemediation() in tests
The existing code was too complex, buggy and only worked by accident.
The primary issue was that the code was testing for the
machinepoolconfig to contain the MC created by the operator even if the
MC was being removed. Another issue was that there were no debug
messages in the tests, which meant that debugging was too difficult.