Refactor: rename variable
I am trying to unbundle MachineConfig API from the Compliance Operator as these
two start to diverge.

First roadblock, I have hit was the McRoleKey not available in the original
MachineConfig. I did submit the variable on the MachineConfig side at
https://github.com/openshift/machine-config-operator/pull/1418
and here I rename the bundled variable in Compliance Operator so we get closer
to the original.