Add prefixes to the scans created by the compliancesuite
In order to make the names unique, this patch adds a prefix to the name
of the compliancescans created by the compliancesuite controller. This
way, we can have multiple suites that create scans with similar names
without conflict.

Note that this keeps the behavior of showing the non-prefixed scan names
in the suite scan status.

This addresses #84