Add a benchmark test
This adds a benchmark test to help us identify the bottleneck in the XML
parser.

Currently the only package ran by the test is the pkg/utils package,
since that's where the bottleneck lives. This is because benchmark tests
can only be ran on one package.

This also introduces the `make test-benchmark` target, which runs the
aforementioned tests.