suite: Wait for CMs a little as a temporary workaround
As we found out through test failures, it is possible that a resource is
created but cannot be found immediatelly afterwards. This was especially
the case with the huge configmaps we create to hold the results.

The proper way of fixing the issue would be to don't exit the
logcollector container until the configmap it creates is really there,
but in order to fix our CI first and unblock other development, let's
just poll for a bit.