Move resultscollector Dockerfile to the top-level directory
This fixes the "make resultscollector-image" target work. Previously,
the Dockerfile was executed in the context of the
"images/resultscollector" directory, so the COPY directive only copied
what was in that directory. And COPY doesn't support navigating outside
the build context.

The reason the target worked at all was that there was a typo in the
RESULTSCOLLECTOR_DOCKERFILE_PATH variable, it was pointing to the
openscap scanner Dockerfile where we don't have the same issue as we
only need the files from the context to build the image.