Wait for ConfigMaps in compliancescan controller
We used to just move to the DONE phase after getting the results from
the openscap container... this wasn't very practical for the components
depending on the results (e,g, the compliancesuite controller). So we
just wait for all of the configmaps to be ready in the AGGREGATING phase
of the compliancescan controller.