Don't upload the ARF to ConfigMaps, instead use the DS and the XCCDF results separately
Changes the shell script in the openscap component so that after the ARF
is generatd, it is also split into the XCCDF and the DS parts. The full
ARF is still being uploaded to the resultsserver, by the logcollector,
but the XCCDF is what's now being uploaded to the configmaps.

Provides a new utility function to parse the XCCDF,DataStream pair and return
remediations out of that. Removes the ARF parsing as it's no longer used,
along with the huge ARF test files.