Add a new component, the CM aggregator
We've been using the full ARF report as produced by the openscap-chroot
tool because it contains both the input DS and the test results.
However, as the DS gets too big, the ARF report would no longer fit into
the 1MB ConfigMap limit.

To address this, we're going to switch to only generating the XCCFD
report out of the ARF report and then generate the remediations based on
a (DataStream, XCCDF) pair. The openscap container would upload not the
ARF, but only the XCCDF which is significantly smaller.

Because the DataStream comes from the init container and generally can be
user-provided, we need a new pod that processes the XCCDF reports together
with the Data Stream. This patch adds the aggregator binary that would be
running in this new pod, but the binary is not used yet as of this
patch.