e2e: Speed up tests by pausing a MachineConfigPool before modifying it
Some tests need to apply multiple MachineConfigs during their run.
Because normally, applying a MC means the cluster gets rebooted,
applying N MCs meant N reboots. If we pause a cluster by setting
paused=true prior to applying a MC and then unpause after all the MCs
are updated, we can only reboot the cluster once during the whole test.

Required a little refactoring of the functions that wait for the
different state changes in the cluster as some of the conditions are
different when a pool is paused, for example we can't check that a MC is
present in a pool's status, but only in its spec.