Optimization: Only parse DS XML once in the aggregator
For each ConfigMap, we were parsing the DS file's XML... Which was quite
costly. Instead, lets parse it once, and reuse the xmldom.Document file
for each configMap, as that file's not modified.

In order to keep the interface parser-agnostic, this introduces a
wrapper to the utils package called XMLDocument.