e2e: Use sub-pool to speed up the e2e test
It is possible to label a node with several pools, creating a sub-pool:
    https://github.com/openshift/machine-config-operator/blob/master/docs/custom-pools.md

We leverage this in our e2e tests in order to create a pool with a
single machine only. This has the benefit of rebooting only one node
when applying or unapplying a remediation.

In my testing with an AWS cluster with three workers, running the tests
that actually touch remediations is about twice as fast using this
optimization.

Because creating or removing a pool also triggers a reboot of the node
that we label, we also add a new structure mcTestCtx that is
instantiated and torn down together with the whole suite. Any pool
created for the e2e test run is tracked by this structure and when the
suite is torn down, the nodes are unlabeled and then the pool is
removed, therefore we only incur the overhead of creating and removing
the pool once per suite run.