Always prepend "/content/" to content path when using it
This used to only prepend "/content" if a relative path was being used.
This remove this "feature" with the intent that "/content" should stay
as an implementation detail and is not something the user needs to know
or care about.

Instead, lets always prepend the "/content" directory, since that's what
we use as a volume mount anyway. The content path will always be
relative to "/", so both relative and absolute paths should keep
working.

Note that the init containers now pipe to /bin/true, this is so they
don't fail, and the failure is instead caught by the commands that would
take the content into use.