Add branch creation as part of the release process
Before, we were just creating a tag and pushing it. While this worked,
we still needed to do a branch and a pull request afterwards...

This fixes that by creating a branch immediately, pushing the tag and
then pushing the branch. At the end of the release target run, you
should get a link to create a pull request.