Make csv target idempotent
As a Makefile would usually have as a good practice... if a directory or
file already exists, you probably don't need to try the command again
and again. So this makes the csv command idempotent by checking that the
directory of the target version already exists.

This also adds docs for the target, so it appears in the help target
too.