Implement autoApplyRemediations
This implements the feature to auto apply remediations by checking if
the scans are all done and then starting to set the "apply" for all of
the remediations belonging to the suite.

Before applying, this actually gets the MachineConfigPool that the
remediation will apply to, and it pauses it. Once all of the "apply"
have been set, it unpauses the pool.

Note that this is only done for Remediations that are of type
MachineConfig.

This also removes the check for nodeSelectors in the suite. This is
instead done on the remediation controller and only for MachineConfig
remediation types.