suite controller: Expose suite copy on reconciler functions
Instead of making a copy in each function, this makes a copy in the
beginning and passes it on to the reconciler, this exposes changes in an
easier manner than having to refetch the suite. However, a copy is still
needed before each udpate, since that function adds metadata to the
object. Copying removes that metadata and gives us a fresh object we can
update with.