Makefile: Use `oc apply` instead of `oc create`
This makes the logic easier, by always ensuring that the resource is in
the cluster, instead of trying to create it and ignoring errors. This
way we'll see less noise in the logs as well as better error handling
(not real errors will make this fail, instead of always ignoring it).