Do creator functions for wrapper-types
Currently, there are two ways of handling wrapper types (which are used
in the complianceSuite):

* From object to wrapper
* From wrapper to object

The way that this is done currently is by setting each value from an
object to another.

This is error prone as it can lead to us forgetting to set some
attribute, specially as the types evolve.

This patch instead creates functions to transform between objects and
their wrappers, and attempts to do it in a way that includes bigger
pieces of the object, e.g. setting the whole ComplianceScan Spec, from
the wrapper's Spec. In the aforementioned example, every time we include
new parameters in the ComplianceScan's spec, the function will
automatically take it into account, so we'll have less code changes
there.