Remove usages of mco client
As opposed to using generated clients, dynamic clients are encouraged
amongst the community. So, instead of relying on a generated client that
might or might not stay in that repo, lets just use our already existing
dynamic client coming from the operator-sdk.

The e2e changes try to be the least intrusive as possible, so the
"mcTestCtx" object is left as-is, and the logic maintainted too.
However, the mco-client is removed from there.