suite: Add aggregated phase and result
In order to view a cummulative phase and result of a phase, this patch
extends the complianceSuite CR with two new attributes:
    aggregatedPhase
    aggregatedResult

With both of them, the logic is similar in the sense that the lowest
phase is always displayed, for example if all scans are already done
except one that is still running, the aggregated phase would be running.

The result is a bit more tricky because there is no natural ordering, so
there the most 'severe' result is displayed.

Also extends the tests to make use of this new phase and result to check
the suite result.