Improve operator logging
The operator was sometimes logging too much and sometimes too little.
This patch adds more log messages to error conditions as well as
branches that were not decorated with logging at all such as pausing or
unpausing the pools.

In addition, some log messages are not completely removed but instead
commented out for when we introduce log levels that let us fine-tune the
verbosity.