Remove remediations from suite status
this is just causing a lot of unnecessary hassle instead of helping...

There are many remediations to keep track of and could potentially
damage the overall performance, since tracking big objects is not ideal
in k8s.

So lets just make use of the labels to get the appropriate remediations.