Remediations: Only update merged MC if remediation isn't there yet
We used to always update the merged MC, however, this was resulting in
multiple changes to the MC because the reconciler would be called many
times and the updates aren't ordered (we can't asure that). So, to avoid
this, we only update if the remediation isn't already there.

We detect that by checking a specific entry in the MC's annotations.