Generic Remediations round #2
This implements the missing bits to handle generic remediations in the
operator. If the remediation is not a MachineConfig, it'll just ensure
that the object exists in the cluster.

If the remediation is of an object of a kind that doesn't exist in the
cluster, the error gets reflected in the Remediation CR.