Make ComplianceScan re-scan if an annotation is set
This makes the annotation `compliance.openshift.io/rescan` execute a
rescan if it's encountered. Note that this only happens in the "DONE"
phase, and if it's found, it also deletes the previous pods and secrets
before going back to the "PENDING" phase. This is done in order to start
from a blank slate and avoid errors.

On the other hand, this introduces the `currentIndex` field to the
ComplianceScan's status. This represents the number of scans that have
been executed, and serves as a directory for the results in the PVC.