Replace explicit object additions for AddToScheme command
We don't need to add each object to the scheme... we can just do it in
one go.

I only added this to the common.go file in cmd/managed since this will
be the one that we'll use once we move the rest of the subcommands to a
single binary.