Surface validation error to ComplianceSuite object
If the ComplianceSuite object has validation errors (e.g. the schedule
having the wrong format), instead of continuing without failing, this
makes it so that it won't schedule the ComplianceScan(s) and it'll
surface the error immediately.

This error will be shown in the logs, an event and in the
ComplianceSuite's status.