Decouple resources from CRs
This decouples the infrastructure (all the resources that we create
in order to run scans) to the CRs. The intent is to be able to
create scans, suites and their results in arbitrary namespaces,
while the infra will stay in the compliance operator's namespace.

This first run removes all the instances of ownership from the CRs to
the infra and ensures that the infra gets creates in the appropriate
namespace (the operator's namespace).