Make PKI elements to be in the same namespace as the ComplianceScan
With the current architecture we have, we need to have the infra
objects be in the same namespace as the scan/suite. this is because of
our usage of "SetControllerReference" which sets the suite/scan as the
owner of the objects. This doesn't work cross-namespace. So let's ensure
it works.

This way we'll be able to use OLM's 'targetNamespace' capability, which
ensures our operator can listen for objects in other namespaces.