ComplianceSuite: Use finalizers to clean up resources
This adds finalizers to the ComplianceSuite object in order to clean up
its resources as well. Namely the only thing that gets cleaned up by the
finalizer is the CronJob (if it exists). Scans or Remediations should be
in the same namespace and actually retained ownership, so they would get
cleaned up by the garbage collector.