Enable multiple watch-namespaces to be set
This enables the operator to use multiple watch-namespaces, which can be
given by separating them with the "," separator.

In order to keep the operator working, we overwrite the namespace to
always include the operator's namespace. This way the cache can also
track the resources that the operator takes into use.

This also exposes an environment variable called "OPERATOR_NAMESPACE"
which will be taken into use in a local run.