Standard flags for all subcommands and common way of getting kubeconfig
This enables standard flags for all subcommands (logging configurations
and kubeconfig related flags) which will enable us to configure similar
capabilities for all of them.

With this into use, this also enables a common way of getting the
cluster/client configuration. It can now dynamically check if it's in
cluster environment or if it needs to use a kubeconfig, thus enabling us
to use all of the subcommands locally if needed.