Label Profiles and Rules; and test
In order to easily differentiate between what ProfileBundle owns what
profiles and rules, a label has been created.

This is tested in an e2e test that tests both that the profiles and
rules were correctly parsed (the data stream being correct) and that the
objects were labeled and we can filter them.