Add a ScanSettingBinding controller
Introduces two new CRs: ScanSettingBinding and ScanSetting.
ScanSettingBinding is the one the controller watches as primary. It
contains references to Profiles or TailoredProfiles as well as a
reference to ScanSetting. Based on information in these CRs, the
controller generates a ComplianceSuite without the administrator having
to fill in any of the low-level SCAP data like content image or file.

The controller also watches the ScanSetting as a secondary resource and
maps updates to the Setting CRs to Reconcile events of any
ScanSettingBinding that uses this setting in order to allow updates.