Fix race condition when dealing with missing tailoring config map
Previously, we set the scan's result to error if the configmap wasn't
found, then requeued in and updated.

This was problematic as at points we would get an error, and at other
points we would get the desired result. Instead, here we clean up the
error status if the configmap is subsequently found.

We also don't do status updates unnecessarily, only issue events.