Debug flag no longer cleans up resources
While at first glance it seemed like a good idea. It turns out that it's
useful to keep resources around to be able to debug and diagnose
deployments.

So, this changes the debug flag to no longer clean up pods and
resources. Instead it just enables debug logging of workloads and
OpenSCAP.