Handle resource limit issues when allocating raw result storage
It may be the case that there are restrictions in the cluster that
should be taken into account (e.g. via the usage of LimitRanges or
ResourceQuotas). This would result in authorization errors when creating
the PVC that we use to allocate raw storage.

This takes that into account by detecting this and outputting an error
up to the ComplianceScan/Suite level.