Enable raw result rotation
This adds a flag called `rawResultStorageRotation` which, allows users to
configure the rotation for the result directories in the PVC.

It defaults to '3' and will skip the `lost+found` directory if it's
present.

It works by checking the creation date for the directories and making
the rotation decision based on that. This way we don't have to think
about the indexing names and rotation of indexes.

A rotation policy of 0 disables the rotation entirely.