Only encode to base64 results compressed in configMaps
Base64 encoding results, while it's useful for URLs and things that need
to be URL-encoded, it takes more space ultimately, which results in more
bytes sent through the network.

To avoid this, we don't do this when sending them to the result server.
the result server instead will just get the bzip2 files, and we only
base64 them for ConfigMaps.