aggregator: Don't fail if remediation cannot be created
In cases where remediations can't be created (we don't know what role to
use), we shouldn't fail but instead warn and issue a message about it.

This skips the offending remediation and continues creating the check
results.

When a remediation cannot be created, an event is issued on the
ComplianceScan object.