Encapsulate RawResultStorage options into its own section
This introduces a section to the scan called `RawResultStorage`, this
sets storage related options through the RawResultStorageSettings
object.

As we add more storage-related options to these objects, this makes sure
that we don't use that many redundant names (e.g. RawResultStorage as a
prefix for all of them) and instead have them all in one logical place.