make: Point out the required courier version
Both operator-courier version as shipped in Fedora and the latest
upstream version have issues validating CRDs that were generated using
recent operator-sdk versions, just different ones. The version as
shipped in Fedora (2.1.7) has issues validating CSVs that are generated
from newer CRD versions and the versions 2.1.8 that was supposed to fix
this has issues matching the CRD version with CSV versions. Luckily a
patch is available upstream to fix the latter issue, so a combination of
2.1.8 and that patch works fine.

This commit just adds a pointer to what is needed along with a pointer to
a COPR repo with a patched 2.1.8 courier.

Jira: CMP-683