make: Fix detecting the previous version
Our detection of the latest released version was too simplistic in the
sense that it was just using alphabetical sort, so 0.9 evaluated as
higher than 0.10.

This patch adds a python script that is also not exactly the most
complete semver parser, but works OK. Our use-case is narrower anyway,
we don't bother with equal versions (we only try to find the highest)
and we know that the versions are always major.minor.patch as that's
what is supported by operator-sdk.

The patch also moves the evaluation of the OLD_COURIER_PACKAGE_VERSION
to the only Makefile rule that needs it, otherwise any make invocation
would have exec-ed python.

Jira: CMP-683