types: Split a Payload inlined structure from Rule,Variable and Payload types
We're going to want to update the Rule, Variable and Payload types.
Normally, this is done by updating either the Spec or the Status while
retaining the data in ObjectMeta and TypeMeta, but these structs had
only direct attributes, not a Spec or a Status.

To allow the updates easily and sort of simulate the easy update
behaviour, this patch splits the the user-settable data into a Payload
structure that can be updated on its own.

Jira: CMP-665