profileparser: Update data parsed out from a ProfileBundle
In order to provide the most up-to-date view of data parsed from a
ProfileBundle, the PB controller already re-runs a pod each time a new
image is detected.

In addition, this patch updates objects parsed from the bundle such as
Profiles, Rules and Variables. After the existing objects have been
updated, all objects that are not parsed from the current bundle are
removed. The detection is implemented by annotating the objects with the
current image digest, so the cleanup removes all objects that are not
annotated with that digest.

The code that parses the whole bundle has been moved from the manager to
the profileparser package so that it is easier to test. Most of the
tests are unit tests, the e2e tests are a subset of the unit tests.

Jira: CMP-665