Allow updates of compliancescans by modifying compliancesuite
Previously, once ComplianceSuite was created and spawned scans, that was
it, the scans could only be modified by re-creating the suite.  This patch
allows modifications of scan CR if the scan spec in the suite is modified.

The scan can only be modified in DONE, because otherwise consistency is
hard to guarantee.

JIRA: CMP-679