Alow update of MC remediations
If contents of remediation that was previously aplied is changed, the
remediation is now switched to a new state called "Outdated". The
previous contents are stored in a separate attribute and the admin is
supposed to choose between the outdated or the new content, IOW the
admin would accept the new remediation contents by removing the outdated
fields.

The outdated remediations are also labeled and a suite event is issued when
any such remediations exist.

Jira: CMP-630