Move CM creation to launching phase
It used to be in the pending phase. However, in cases where the CM would
not be created or would be deleted half way, we would end up in an
unrecoverable state.

This moves the creation of CMs to the launching phase which makes it
happen in the same state where the pods are created. If, for some
reason, the CM would be deleted while the pods are launching, this gives
us a way to recover (by deleting the pod, which will move the state from
running back to launching).

The pending state is now just for validations.