images: Use a fallback RELATED_IMAGE_OPENSCAP value that actually works
We recently changed the environment variable names to match downstream,
which caused in one instance the operator to be assigned the new names,
while the code was still handling the old names.

This uncovered the issue this patch is trying to address: the fallback
image for when RELATED_IMAGE_OPENSCAP is not set is too old and it was
causing the platform scans to fail because the old scap version doesn't
understand the YAML probes.

Jira: OCPBUGSM-13604