Refactor and validate scan types
This now validates the scan type at the beginning of the run (on the
pending phase). If it's not a valid type it'll return an error.

This would work since now we default (via the CRD) to Node.

Validation is also done so the types are not case-sensitive (to improve
UX).

We now can use the types through the code base instead of the custom
validation we had.