Detect if scan is unscheludable
If a pod (or more) is not schedulable in the scan due to a taint or
another scheduling issue, we detect this and issue partial results;
however, an error is persisted as the ultimate result of the scan.

This is done by creating a configMap (as the log-collector would do)
that contains a status that the node didn't get a scan due to it being
unschedulable, and adding the pod's condition as an error message.

This also changes how the aggregator is scheduled, as it now only gets
skipped if there's an unknown error.