build/Dockerfile: Don't create files as root when building with docker
Turns out that podman is not 1:1 compatible with docker, but in this
case it's a good thing, because docker's behaviour is just plain
dangerous - even if a container is running as a non-privileged user by
default (like, the UBI image runs as the user "default"), they'd still
make the files copied with the COPY directive owned by root. Because
honoring the permissions would be a breaking change, docker instead
introduced a --chown parameter to COPY:
    https://github.com/moby/moby/issues/13020
    https://stackoverflow.com/questions/44766665/how-do-i-docker-copy-as-non-root
which podman luckily supports as well.

Without this patch, creating the "_output" directory during the build
would have failed with a Permission denied.