make: Don't use podman-specific semantics of the build -f flag
This patch fixes a podman-specific invocation in the Makefile: podman
allows to start a build by pointing at a Dockerfile with the -f flag and
the directory that contains the Dockerfile then becomes the build
context.

With docker, this is not supported, one must instead point to the
context as a positional argument and then the Dockerfile is used
automatically from there.

This change is forward-compatible with podman.