Make access modes and storage class configurable for PVs
In cases where there isn't a default storage class configured or simply
the admin wants to use something other than the default, we are now able
to configure the storage class through the raw result storage settings
in the scan/suite.

This requires us to be able to set the access modes, which could vary
between storage classes. By making this configurable, this gives the
advantage of allowing admins to have a long-running process that
forwards the raw results if the storage class supports a ReadWriteMany
access mode. Note that it would still be easier to just clone the
volume...