Rename aggregatedPhase and aggregatedResult to just Phase and Result
The OCP console currently does not honor additionalPrinterColumns.
Instead, the console looks for one of hardcoded attributes:
    status.phase
    status.status
    status.state
    status.conditions
to support displaying the suite status in the same way we display the
Scan status - for scans we already have status.phase, for suites we used
to have status.aggregatedPhase. This patch uses status.phase for the
phase aggregated from scans to suite.

Jira: OCPBUGSM-8550