Simplify README
The README now only contains a small walkthrough to show folks how to
use the operators.

This walkthrough takes into use the scansettingbindings object which
uses a default profile (rhcos4-moderate) and the scansettings object
we ship by default.

On the other hand, this moved the CRD descriptions to its own doc.

Documenting how to create tailored profiles will come as a subsequent
PR.