Workaround: Only use profileparser as an init container
The profileparser pod takes a lot of resources and doesn't let them
go... This shows up as excessive resource usage in the monitoring
dashboard (up to 400MiB).

By changing that profileparser to be an init container, as opposed to a
daemon, the parser process will end and the resources will be freed. The
daemon bit will now be a "pauser" container, which merely waits for a
kill signal. This way, this container takes a very minimal amount of
resources and it keeps running as is needed by the kubernetes' Deployment
that manages it.