utils: Don't use ctime but mtime instead
ctime updates even when the contents of the files don't update, e.g.
with permissions update or adding links. It appers that when the
resultserver is scaled up or down, the ctime is always updated,
resulting in the sort of the directories being quite random.

Let's use mtime instead.

Jira: OCPBUGSM-13482