MachineConfig: Use generation to check if we have an up-to-date remediation
Our previous method of checking if machineconfig remediations were up to
date had a flaw: It was order dependant. This means that whenver the
merging of machine configs from remediation ended up in a different
order (e.g. fore files or units) we would have a differing machine
config, and so MCO would create a new rendering.

This would end up in an endless loop and thus the cluster would keep
restarting.

This changes the way we check for remediations being up-to-date, by
using the remediation's generation. If there is a change in the
remediation definition, then we would see a generation change. Else, we
assume that all is good and up to date.